# backend/openapi.yaml
openapi: 3.1.0
info:
  title: tfbpshiny-go API
  version: "1.0.0"
  description: |
    HTTP/JSON API for the Go rewrite of tfbpshiny. The contract is frozen at the
    end of Phase 1; the React SPA (Phase 2) generates TypeScript types from this
    document via `openapi-typescript`.

    Every `/api/v/{v}/*` endpoint compares the path-embedded artifact version
    against the running artifact and responds with `410 Gone` + a `Location:
    /api/version` header when they disagree. The frontend uses this to detect
    stale-tab scenarios and force a refetch of `/api/version`.

servers:
  - url: /
    description: Same-origin (Vite dev proxy in development; Traefik in production)

tags:
  - name: ops
    description: Liveness, readiness, version, and metrics endpoints
  - name: catalog
    description: Dataset and regulator metadata
  - name: data
    description: Binding and perturbation data queries
  - name: comparison
    description: Binding x perturbation comparison endpoints

paths:
  /healthz:
    get:
      tags: [ops]
      summary: Liveness probe
      description: Always returns 200 if the process is running.
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthzResponse"

  /readyz:
    get:
      tags: [ops]
      summary: Readiness probe
      description: |
        Returns 200 when the DuckDB pool is open, the artifact_manifest canary
        query succeeds, and the cache is initialized. Returns 503 with a
        `reason` field otherwise.
      responses:
        "200":
          description: Service is ready to accept traffic.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyzResponse"
        "503":
          description: Service is not ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyzResponse"

  /api/version:
    get:
      tags: [ops]
      summary: Running artifact version metadata
      description: |
        Returns the currently-loaded artifact's version, schema version, build
        timestamp, and DuckDB version. The `artifactVersion` value is the
        `{v}` path segment all `/api/v/{v}/*` endpoints expect.
      responses:
        "200":
          description: Version metadata for the loaded artifact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionInfo"

  /api/v/{v}/datasets:
    get:
      tags: [catalog]
      summary: List available datasets and their fields
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
      responses:
        "200":
          description: All datasets registered in the artifact's dataset_manifest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetsResponse"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/regulators:
    get:
      tags: [catalog]
      summary: Search regulators by locus tag or symbol
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: search
          in: query
          required: false
          description: Case-insensitive prefix match against locus_tag and symbol. Capped at 64 chars.
          schema:
            type: string
            maxLength: 64
        - name: limit
          in: query
          required: false
          description: Maximum rows returned. Clamped to [1, 100]; defaults to 25.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Matching regulators.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegulatorsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/regulators/resolve:
    get:
      tags: [catalog]
      summary: Resolve a user-supplied regulator list to canonical locus tags
      description: |
        Added in Phase 2 Task 2. Accepts a list of free-form regulator tokens
        (locus tags, symbols, or display names) and returns the canonical
        locus_tag list. Supports a `common` mode that intersects against the
        set of regulators present across requested datasets.
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: regulators
          in: query
          required: false
          description: |
            Comma-separated list of regulator locus tags to resolve. Capped at
            30 raw entries (request rejected if exceeded); the server deduplicates
            and uppercases entries before processing.
          schema:
            type: string
        - name: common
          in: query
          required: false
          description: |
            Sugar for `intersect`: a colon-separated pair `A:B` of dataset
            db_names (e.g. `callingcards:hackett`). Equivalent to
            `intersect=A,B`. Optional `binding.` / `perturbation.` prefixes are
            accepted on each side. If both `common` and `intersect` are set,
            `common` wins.
          schema:
            type: string
        - name: intersect
          in: query
          required: false
          description: |
            Comma-separated dataset db_names to intersect. Each name must be in
            `dataset_manifest`. Optional `binding.` / `perturbation.` prefixes
            are accepted; the prefix (if present) must match the dataset's
            actual `data_type`. The list is capped at the dataset manifest size.
          schema:
            type: string
      responses:
        "200":
          description: Canonical locus_tag list (deduplicated, order preserved).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/binding:
    get:
      tags: [data]
      summary: Binding rows for one regulator across one or more binding datasets
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: regulator
          in: query
          required: true
          description: Regulator locus_tag (required).
          schema:
            type: string
        - name: datasets
          in: query
          required: false
          description: |
            Comma-separated binding dataset db_names. Each must be present in
            `dataset_manifest` AND have `data_type=binding`. Limited to the
            count of datasets in the manifest.
          schema:
            type: string
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: Per-dataset binding rows for the regulator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BindingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/perturbation:
    get:
      tags: [data]
      summary: Perturbation rows for one regulator across one or more perturbation datasets
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: regulator
          in: query
          required: true
          description: Regulator locus_tag (required).
          schema:
            type: string
        - name: datasets
          in: query
          required: false
          description: |
            Comma-separated perturbation dataset db_names. Each must be present
            in `dataset_manifest` AND have `data_type=perturbation`. Limited to
            the count of datasets in the manifest.
          schema:
            type: string
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: Per-dataset perturbation rows for the regulator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerturbationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/binding/corr:
    get:
      tags: [data]
      summary: Per-regulator correlation between two or more binding datasets
      description: |
        Returns Pearson- or Spearman-correlated (regulator, sample_a, sample_b)
        groups for every pair drawn from `sorted(datasets) choose 2`. The
        underlying SQL templates (`backend/internal/queries/binding/corr_pair_*.sql`)
        mirror Shiny's `_corr_pair_sql_impl` exactly: INNER JOIN on (regulator,
        target), NULL/Inf/NaN exclusion, and a `HAVING COUNT(*) >= 3` floor.
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: datasets
          in: query
          required: true
          description: |
            Comma-separated binding dataset db_names. At least 2 entries; all
            must be present in `dataset_manifest` AND have `data_type=binding`.
          schema:
            type: string
        - name: method
          in: query
          required: true
          description: Correlation method (`pearson` for raw values, `spearman` for ranks).
          schema:
            type: string
            enum: [pearson, spearman]
        - name: col
          in: query
          required: true
          description: |
            Measurement column kind. `effect` selects `dataset_manifest.effect_col`;
            `pvalue` selects `dataset_manifest.pvalue_col` (falling back to
            `effect_col` when the dataset has no p-value column, matching Shiny's
            `get_measurement_column`).
          schema:
            type: string
            enum: [effect, pvalue]
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: Correlation envelope; one CorrPair per `(dbA, dbB)` combination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/binding/scatter:
    get:
      tags: [data]
      summary: Per-target value pairs for one regulator across two binding datasets
      description: |
        Returns (target, val_a, val_b) tuples for a single regulator across one
        pair of binding datasets. For `method=spearman`, val_* are rank values
        produced by `RANK() OVER`. The server computes `r` (Pearson correlation
        over the returned points) using the same algorithm class as DuckDB's
        `corr()`; the value is clamped to [-1, 1] and coerced to 0 when fewer
        than 2 points are returned or the variance is zero (mirrors Shiny's
        NaN-coercion behavior).
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: regulator
          in: query
          required: true
          description: Regulator locus_tag (required); bound as a positional `?` in the SQL.
          schema:
            type: string
        - name: pair
          in: query
          required: true
          description: |
            Comma-separated pair of binding dataset db_names (exactly 2 entries).
            Both must have `data_type=binding`. Order is preserved (dbA, dbB).
          schema:
            type: string
        - name: method
          in: query
          required: true
          schema:
            type: string
            enum: [pearson, spearman]
        - name: col
          in: query
          required: true
          schema:
            type: string
            enum: [effect, pvalue]
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: Per-target value pairs for the regulator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScatterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/perturbation/correlations:
    get:
      tags: [data]
      summary: Per-regulator correlation between two or more perturbation datasets
      description: |
        Perturbation analogue of `/binding/corr`. The path uses the plural
        `correlations` (vs `corr` on the binding side) to match the original
        Shiny module's naming and the parity audit in
        `docs/parity/perturbation.md`. Same response shape as `/binding/corr`.
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: datasets
          in: query
          required: true
          description: |
            Comma-separated perturbation dataset db_names. At least 2 entries;
            all must have `data_type=perturbation`.
          schema:
            type: string
        - name: method
          in: query
          required: true
          schema:
            type: string
            enum: [pearson, spearman]
        - name: col
          in: query
          required: true
          schema:
            type: string
            enum: [effect, pvalue]
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: Correlation envelope; one CorrPair per `(dbA, dbB)` combination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/perturbation/scatter:
    get:
      tags: [data]
      summary: Per-target value pairs for one regulator across two perturbation datasets
      description: |
        Perturbation analogue of `/binding/scatter`. See that endpoint's
        description for the per-target row shape and server-side Pearson r
        computation.
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: regulator
          in: query
          required: true
          schema:
            type: string
        - name: pair
          in: query
          required: true
          description: Comma-separated pair of perturbation dataset db_names.
          schema:
            type: string
        - name: method
          in: query
          required: true
          schema:
            type: string
            enum: [pearson, spearman]
        - name: col
          in: query
          required: true
          schema:
            type: string
            enum: [effect, pvalue]
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: Per-target value pairs for the regulator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScatterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/comparison/topn:
    get:
      tags: [comparison]
      summary: Top-N responsive-ratio rows for binding x perturbation pairs
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
        - name: binding
          in: query
          required: false
          description: Comma-separated binding dataset db_names.
          schema:
            type: string
        - name: perturbation
          in: query
          required: false
          description: Comma-separated perturbation dataset db_names.
          schema:
            type: string
        - name: top_n
          in: query
          required: false
          description: Top N rows per (binding, perturbation) pair. Clamped to [1, 10000]; defaults to 25.
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 25
        - name: effect
          in: query
          required: false
          description: Effect-size threshold for the responsive predicate. Default `0.0`.
          schema:
            type: number
            format: double
            default: 0.0
        - name: pvalue
          in: query
          required: false
          description: P-value threshold for the responsive predicate. Default `0.05`.
          schema:
            type: number
            format: double
            default: 0.05
        - $ref: "#/components/parameters/Filters"
      responses:
        "200":
          description: TopN rows across the requested binding/perturbation pairs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopNResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

  /api/v/{v}/comparison/dto:
    get:
      tags: [comparison]
      summary: Precomputed DTO (downstream/target-overlap) comparison rows
      parameters:
        - $ref: "#/components/parameters/ArtifactVersion"
      responses:
        "200":
          description: All rows from the DTO precomputed comparison table.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DTOResponse"
        "410":
          $ref: "#/components/responses/StaleArtifactVersion"

components:
  parameters:
    ArtifactVersion:
      name: v
      in: path
      required: true
      description: |
        Artifact version (e.g. `2026-05-12.1`). Must equal the running
        artifact's `artifactVersion`; otherwise the endpoint returns 410.
      schema:
        type: string

    Filters:
      name: filters
      in: query
      required: false
      description: |
        URL-encoded JSON object of shape `FiltersByDB` (see schema). The raw
        string is capped at 16 KiB before unmarshal to prevent DoS.
      schema:
        type: string
        maxLength: 16384

  responses:
    BadRequest:
      description: Invalid query parameter or filter shape.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    StaleArtifactVersion:
      description: |
        The `{v}` path segment does not match the running artifact. Clients
        should refetch `/api/version` (advertised via the `Location` header)
        and retry against the new version.
      headers:
        Location:
          description: Always `/api/version`.
          schema:
            type: string
            example: /api/version
      content:
        text/plain:
          schema:
            type: string
            example: stale artifact version

  schemas:
    HealthzResponse:
      type: object
      required: [alive]
      properties:
        alive:
          type: boolean
          example: true

    ReadyzResponse:
      type: object
      required: [ready]
      properties:
        ready:
          type: boolean
        reason:
          type: string
          description: Present only when `ready` is false.

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.

    VersionInfo:
      type: object
      required: [artifactVersion, schemaVersion, builtAt, duckdbVersion]
      properties:
        artifactVersion:
          type: string
          description: Artifact version baked into every cache key.
        schemaVersion:
          type: integer
          description: Schema version compatible with this server build.
        builtAt:
          type: string
          format: date-time
          description: RFC3339 timestamp of when the artifact was built.
        duckdbVersion:
          type: string
          description: DuckDB storage version (informational only; not a startup gate).

    DatasetEntry:
      type: object
      required: [dbName, dataType, assay, displayName, sourceRepo, sampleIdField, fields]
      properties:
        dbName:
          type: string
        dataType:
          type: string
          description: Either `binding` or `perturbation`.
        assay:
          type: string
        displayName:
          type: string
        sourceRepo:
          type: string
        sampleIdField:
          type: string
        fields:
          type: array
          items:
            type: string

    DatasetsResponse:
      type: object
      required: [datasets]
      properties:
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/DatasetEntry"

    Regulator:
      type: object
      required: [locusTag, symbol, displayName]
      properties:
        locusTag:
          type: string
        symbol:
          type: string
        displayName:
          type: string

    RegulatorsResponse:
      type: object
      required: [regulators]
      properties:
        regulators:
          type: array
          items:
            $ref: "#/components/schemas/Regulator"

    ResolveResponse:
      type: object
      required: [regulators]
      properties:
        regulators:
          type: array
          description: |
            Canonical, sorted, deduplicated locus_tag list. Capped at 1000
            entries server-side; `truncated:true` indicates the cap was hit.
          items:
            type: string
        truncated:
          type: boolean
          description: |
            True when the resolved list exceeded the 1000-tag cap and was
            trimmed. Clients should consider narrowing the dataset intersection
            or providing an explicit `regulators` list.

    BindingRow:
      type: object
      required: [regulatorLocusTag, targetLocusTag, sampleId, value]
      properties:
        regulatorLocusTag:
          type: string
        targetLocusTag:
          type: string
        sampleId:
          type: string
        value:
          type: number
          format: double

    BindingDatasetResult:
      type: object
      required: [dbName, column, rows]
      properties:
        dbName:
          type: string
        column:
          type: string
          description: Measurement column projected as `value` for this dataset.
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BindingRow"

    BindingResponse:
      type: object
      required: [regulator, datasets]
      properties:
        regulator:
          type: string
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/BindingDatasetResult"

    PerturbationRow:
      type: object
      required: [regulatorLocusTag, targetLocusTag, sampleId, value]
      properties:
        regulatorLocusTag:
          type: string
        targetLocusTag:
          type: string
        sampleId:
          type: string
        value:
          type: number
          format: double

    PerturbationDatasetResult:
      type: object
      required: [dbName, column, rows]
      properties:
        dbName:
          type: string
        column:
          type: string
          description: Measurement column projected as `value` for this dataset.
        rows:
          type: array
          items:
            $ref: "#/components/schemas/PerturbationRow"

    PerturbationResponse:
      type: object
      required: [regulator, datasets]
      properties:
        regulator:
          type: string
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/PerturbationDatasetResult"

    TopNRow:
      type: object
      required:
        - pairKey
        - bindingSampleId
        - regulatorLocusTag
        - perturbationSampleId
        - n
        - nResponsive
        - responsiveRatio
      properties:
        pairKey:
          type: string
          description: Composite key of the form `{bindingDB}__{perturbationDB}`.
        bindingSampleId:
          type: string
        regulatorLocusTag:
          type: string
        perturbationSampleId:
          type: string
        n:
          type: integer
          format: int64
          description: Number of binding-ranked targets considered.
        nResponsive:
          type: integer
          format: int64
          description: How many of those targets cleared the responsive predicate.
        responsiveRatio:
          type: number
          format: double

    TopNResponse:
      type: object
      required: [topN, effectThreshold, pvalueThreshold, rows]
      properties:
        topN:
          type: integer
        effectThreshold:
          type: number
          format: double
        pvalueThreshold:
          type: number
          format: double
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TopNRow"

    DTORow:
      type: object
      required:
        - bindingIdSource
        - perturbationIdSource
        - dtoEmpiricalPvalue
        - dtoFdr
        - bindingSetSize
        - perturbationSetSize
        - bindingSampleId
        - pertSampleId
        - time
      properties:
        bindingIdSource:
          type: string
        perturbationIdSource:
          type: string
        dtoEmpiricalPvalue:
          type: number
          format: double
        dtoFdr:
          type: number
          format: double
        bindingSetSize:
          type: integer
          format: int64
        perturbationSetSize:
          type: integer
          format: int64
        bindingSampleId:
          type: string
        pertSampleId:
          type: string
        time:
          type: string

    DTOResponse:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/DTORow"

    CorrPairPoint:
      type: object
      required: [dbA, dbAId, dbB, dbBId, regulatorLocusTag, correlation]
      properties:
        dbA:
          type: string
          description: Dataset db_name on the A side.
        dbAId:
          type: string
          description: sample_id from dataset A.
        dbB:
          type: string
        dbBId:
          type: string
        regulatorLocusTag:
          type: string
        correlation:
          type: number
          format: double
          description: |
            corr() value for this (regulator, dbAId, dbBId) group. For Pearson,
            computed over raw measurement values; for Spearman, computed over
            RANK() outputs. Clamped to [-1, 1] in practice.

    CorrPair:
      type: object
      required: [dbA, dbB, colA, colB, points]
      properties:
        dbA:
          type: string
        dbB:
          type: string
        colA:
          type: string
          description: Measurement column actually used on the A side.
        colB:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/CorrPairPoint"

    CorrResponse:
      type: object
      required: [method, col, pairs]
      properties:
        method:
          type: string
          enum: [pearson, spearman]
        col:
          type: string
          enum: [effect, pvalue]
        pairs:
          type: array
          description: |
            One entry per `(dbA, dbB)` combination drawn from
            `sorted(datasets) choose 2`. Ordering is stable across param-order
            permutations so cache values are deterministic.
          items:
            $ref: "#/components/schemas/CorrPair"

    ScatterPoint:
      type: object
      required: [targetLocusTag, valA, valB]
      properties:
        targetLocusTag:
          type: string
        valA:
          type: number
          format: double
        valB:
          type: number
          format: double

    ScatterResponse:
      type: object
      required: [regulator, dbA, dbB, colA, colB, method, r, points]
      properties:
        regulator:
          type: string
        dbA:
          type: string
        dbB:
          type: string
        colA:
          type: string
        colB:
          type: string
        method:
          type: string
          enum: [pearson, spearman]
        r:
          type: number
          format: double
          description: |
            Pearson correlation computed server-side over the returned points.
            Coerced to 0 when fewer than 2 points are returned or when the
            denominator is zero (matches Shiny's NaN-coercion). Clamped to
            [-1, 1].
        points:
          type: array
          items:
            $ref: "#/components/schemas/ScatterPoint"

    FilterSpec:
      type: object
      required: [type, value]
      description: |
        One field-level filter clause. `value` is polymorphic and depends on
        `type`:
          - `categorical`: array of strings (IN-list).
          - `numeric`:     two-element [min, max] number tuple (inclusive).
          - `bool`:        plain boolean.
        The backend validates the shape per `type` at request time and rejects
        mismatches with 400.
      properties:
        type:
          type: string
          enum: [categorical, numeric, bool]
        value:
          oneOf:
            - type: array
              description: Categorical IN-list (used when `type=categorical`).
              items:
                type: string
            - type: array
              description: Numeric inclusive range `[min, max]` (used when `type=numeric`).
              minItems: 2
              maxItems: 2
              items:
                type: number
                format: double
            - type: boolean
              description: Boolean filter (used when `type=bool`).

    FiltersByDB:
      type: object
      description: |
        Filters keyed by dataset db_name, then by field name. Both keys are
        identifier-whitelisted at request time against `dataset_manifest` and
        `field_manifest`. Wire-shape is a URL-encoded JSON string passed via
        the `?filters=` query parameter (max 16 KiB).
      additionalProperties:
        type: object
        additionalProperties:
          $ref: "#/components/schemas/FilterSpec"